projects/wines-api/src/requirements.txt
alembic
uwsgi
fastapi
email-validator
uvicorn
python-multipart
pytest
python-dotenv
supabase
python-decouple
pydantic
pydantic-settings
SQLAlchemy>=2.0.23
uuid
psycopg2-binary>=2.9.3
requests
google-genai
numpy
scipy